Rename Avatars namespace under /Users/Commands to Shared
