+++ application/account-management/Core/Features/Users/Shared/AvatarUpdater.cs
@@ -3,7 +3,7 @@ using Microsoft.Extensions.DependencyInjection;
 using PlatformPlatform.AccountManagement.Features.Users.Domain;
 using PlatformPlatform.SharedKernel.Integrations.BlobStorage;
 
-namespace PlatformPlatform.AccountManagement.Features.Users.Avatars;
+namespace PlatformPlatform.AccountManagement.Features.Users.Shared;
 
 public sealed class AvatarUpdater(IUserRepository userRepository, [FromKeyedServices("avatars-storage")] BlobStorageClient blobStorageClient)
 {
